broker/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
pydantic==2.7.3
httpx==0.27.0
pyyaml==6.0.2
